+++ test-scripts/read-write.lua
@@ -1,0 +1,5 @@
+local contents = "Hello, world!"
+
+remodel.writeFile("temp/foo.txt", contents)
+
+assert(remodel.readFile("temp/foo.txt") == contents)